Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main() {
-    float number;
+    double number;
     printf("You entered: ");
     scanf("%1f",&number);
-    printf("%.41f");
+    printf("%.4f");
     return 0;
 }
